// src/types/mock-data.type.ts
export type MockData = {
  titles: string[];
  descriptions: string[];
  genres: string[];
  rating: string[];
  countComments: string[];
  releases: string[];
  previewVideoLinks: string[];
  videoLinks: string[];
  starring: string[];
  directors: string[];
  users: string[];
  emails: string[];
  avatars: string[];
  posterImages: string[];
  backgroundImages: string[];
  backgroundColors: string[];
}